My second commit

+++ test/test.cpp
@@ -30,9 +30,9 @@ TEST_F(PIDTest, flag) {
  * be within a certain tolerance of the desired velocity
  */
 TEST_F(PIDTest, Compute) {
-  float err_tolerance = 0.01;
-  float des_vel = 10.0;
-  float curr_vel = 1.0;
+  double err_tolerance = 0.01;
+  double des_vel = 10.0;
+  double curr_vel = 1.0;
   curr_vel = obj.final_vel(des_vel, curr_vel);
   EXPECT_NEAR(des_vel, curr_vel, err_tolerance);
 }
@@ -45,9 +45,9 @@ TEST_F(PIDTest, Compute) {
  */
 
 TEST_F(PIDTest, doublet) {
-  float err_tolerance = 0.01;
-  float des_vel = 10.0;
-  float curr_vel = 0.0;
+  double err_tolerance = 0.01;
+  double des_vel = 10.0;
+  double curr_vel = 0.0;
   curr_vel = obj.final_vel(des_vel, curr_vel);
   des_vel = -10;
   curr_vel = obj.final_vel(des_vel, curr_vel);
